Change the spec to skip lock field's length
In certain scenarios, the length of the lock field, might only be
decided after the signing. Assuming a multisig setup where different
parties employ different signature validation algorithms(which is
totally possible on CKB), it will only when we have gathered all the
signatures, will we know the final length of the lock field, and the
total witness length.